feat(frontend): extract LocalExecutorGuide component for devices page
- Extract LocalExecutorGuide as reusable component
- Shows 2 steps: curl install from GitHub, run with env vars
- Clean up unused imports in devices page
